Don't mess with global variable `debug`
`debug` is a table in Lua's standard library
(https://www.lua.org/manual/5.1/manual.html#5.9),
and it is absolutely impossible to kill that by setting it to true or false.
Any Lua code down the road will fail with inexplicable errors.

This commit amends this crucial error, but another one will follow
wrapping the whole selnolig.lua file in one global variable because it
is unnecessary to expose so many global variables to the Lua namespace.